requirements.txt
faker==24.8.0
schedule==1.2.1
pymongo==4.11.1
python-dotenv==1.0.1
faker-commerce
mimesis
pandas